+++ config.json
@@ -1,0 +1,26 @@
+[
+  {
+    "module": "lac",
+    "version": "1.0.0",
+    "batch_size": 20,
+    "queue_size": 20
+  },
+  {
+    "module": "senta_lstm",
+    "version": "1.0.0",
+    "batch_size": 20,
+    "queue_size": 20
+  },
+  {
+    "module": "yolov3_coco2017",
+    "version": "1.0.0",
+    "batch_size": 20,
+    "queue_size": 20
+  },
+  {
+    "module": "faster_rcnn_coco2017",
+    "version": "1.0.0",
+    "batch_size": 20,
+    "queue_size": 20
+  }
+]
