+++ paddlehub/serving/app.py
@@ -3,59 +3,159 @@ from flask import Flask, request, render_template
 from paddlehub.serving.model_service.text_model_service import TextModelService
 from paddlehub.serving.model_service.image_model_service import ImageModelService
 from paddlehub.serving import utils
-# from model_service.text_model_service import TextModelService
-# from model_service.image_model_service import ImageModelService
-import json
-import hashlib
 import time
 import os
 import base64
 import logging
-import sys
-import Queue
-import threading
+import cv2
+import multiprocessing as mp
+from multiprocessing.managers import BaseManager
+import random
+import six
+if six.PY2:
+    from Queue import PriorityQueue
+if six.PY3:
+    from queue import PriorityQueue
 
-use_gpu = False
-waiting_queue = Queue.Queue(maxsize=10)
+
+class MyPriorityQueue(PriorityQueue):
+    def get_attribute(self, name):
+        return getattr(self, name)
 
 
-class TestThread(threading.Thread):
-    def __init__(self, thread_id, name, counter):
-        threading.Thread.__init__(self)
-        self.thread_id = thread_id
-        self.name = name
-        self.counter = counter
-        self.g = ""
+class Manager(BaseManager):
+    pass
 
-    def run(self):
-        print("Starting")
-        print("exciting")
-        print("开始阻塞")
-        time.sleep(10)
-        self.g = time.time()
+
+Manager.register("get_priorityQueue", MyPriorityQueue)
+
+
+def choose_module_category(input_data, module_name, batch_size=1):
+    global nlp_module, cv_module
+    if module_name in nlp_module:
+        predict_nlp(input_data, module_name, batch_size)
+    elif module_name in cv_module:
+        predict_cv(input_data, module_name, batch_size)
+
+
+def predict_nlp(input_data, module_name, batch_size=1):
+    global use_gpu
+    real_input_data = []
+    for index in range(len(input_data)):
+        real_input_data.append(input_data[index][3])
+    module = TextModelService.get_module(module_name)
+    method_name = module.desc.attr.map.data['default_signature'].s
+    if method_name != "":
+        predict_method = getattr(module, method_name)
+        try:
+            real_input_data = {"text": real_input_data}
+            results = predict_method(data=real_input_data, use_gpu=use_gpu)
+        except Exception as err:
+            return {"result": "Please check data format!"}
+    else:
+        results = "Module {} can't be use for predicting.".format(module_name)
+    try:
+        result_data = []
+        for index in range(len(input_data)):
+            result_data.append(list(input_data[index]))
+            result_data[-1][3] = results[index]
+    except Exception as err:
+        print("Transform error!")
+    for index in range(len(result_data)):
+        if results_dict.get(result_data[index][2]) is None:
+            results_dict[result_data[index][2]] = [[
+                result_data[index][1], result_data[index][3]
+            ]]
+        else:
+            temp_list = results_dict[result_data[index][2]]
+            temp_list.append([result_data[index][1], result_data[index][3]])
+            results_dict[result_data[index][2]] = temp_list
+    return {"result": results_dict}
+
+
+def predict_cv(input_data, module_name, batch_size=1):
+    global use_gpu
+    filename_list = []
+    for index in range(len(input_data)):
+        filename_list.append(input_data[index][3])
+        cv2.imread(input_data[index][3])
+    input_images = {"image": filename_list}
+    module = ImageModelService.get_module(module_name)
+    method_name = module.desc.attr.map.data['default_signature'].s
+    if method_name != "":
+        predict_method = getattr(module, method_name)
+        try:
+            results = predict_method(
+                data={"image": filename_list}, use_gpu=use_gpu)
+        except Exception as err:
+            return {"result": "Please check data format!"}
+    else:
+        results = "Module {} can't be use for predicting.".format(module_name)
+    try:
+        result_data = []
+        for index in range(len(input_data)):
+            result_data.append(list(input_data[index]))
+            result_data[-1][3] = results[index]
+    except Exception as err:
+        print("Transform error!")
+    for index in range(len(result_data)):
+        if results_dict.get(result_data[index][2]) is None:
+            results_dict[result_data[index][2]] = [[
+                result_data[index][1], result_data[index][3]
+            ]]
+        else:
+            temp_list = results_dict[result_data[index][2]]
+            temp_list.append([result_data[index][1], result_data[index][3]])
+            results_dict[result_data[index][2]] = temp_list
+    return {"result": results}
+
+
+def worker():
+    global batch_size_list, name_list, queue_name_list, cv_module
+    latest_num = random.randrange(0, len(queue_name_list))
+    while True:
+        time.sleep(0.01)
+        for index in range(len(queue_name_list)):
+            while queues_dict[queue_name_list[latest_num]].empty() is not True:
+                input_data = []
+                lock.acquire()
+                try:
+                    for index2 in range(batch_size_list[latest_num]):
+                        if queues_dict[
+                                queue_name_list[latest_num]].empty() is True:
+                            break
+                        input_data.append(
+                            queues_dict[queue_name_list[latest_num]].get())
+                finally:
+                    lock.release()
+                if len(input_data) != 0:
+                    choose_module_category(input_data,
+                                           queue_name_list[latest_num],
+                                           batch_size_list[latest_num])
+                else:
+                    pass
+            latest_num = (latest_num + 1) % len(queue_name_list)
+
+
+def init_pool(l):
+    global lock
+    lock = l
 
 
 def create_app():
-    # app_instance = Flask("paddlehub.serving.app")
     app_instance = Flask(__name__)
     app_instance.config["JSON_AS_ASCII"] = False
     gunicorn_logger = logging.getLogger('gunicorn.error')
     app_instance.logger.handlers = gunicorn_logger.handlers
     app_instance.logger.setLevel(gunicorn_logger.level)
-    global waiting_queue
-
-    @app_instance.route("/test", methods=["GET", "POST"])
-    def test_queue():
-        # print("进程为:", os.getpid())
-        # thread = TestThread(time.time(), "Thread-1", 1)
-        # thread.start()
-        # thread.join()
-        # return {"result": thread.g}
-        print("都可以到这一步")
-        while True:
-            pass
-        global waiting_queue
-        waiting_queue.put(time.time())
+    global queues_dict
+    lock = mp.Lock()
+    pool = mp.Pool(
+        processes=(mp.cpu_count() - 1),
+        initializer=init_pool,
+        initargs=(lock, ))
+    for i in range(mp.cpu_count() - 1):
+        pool.apply_async(worker)
 
     @app_instance.route("/", methods=["GET", "POST"])
     def index():
@@ -63,194 +163,167 @@ def create_app():
 
     @app_instance.before_request
     def before_request():
-        print("start, time= ", time.time())
-        # global waiting_queue
-        print(request.url)
-        print(request.path)
-        data = "data"
-        try:
-            waiting_queue.put(data, block=False)
-        except Queue.Full:
-            return {"result": "现在使用serving的人太多了，请稍后再尝试，谢谢！🙏"}
-        print(request.form)
-        print(request.data)
-        request.data = {"id": "test_id"}
-        # print("正常进行")
-
-    # @app_instance.route("/predict/image/<module_name>", methods=["POST"])
-    # def predict_iamge(module_name):
+        request.data = {"id": str(time.time())}
+        pass
 
     @app_instance.route("/predict/image/<module_name>", methods=["POST"])
     def predict_iamge(module_name):
-        print("123123")
-        print(request.form)
-        print(request.data)
-        global use_gpu
+        global results_dict
+        req_id = request.data.get("id")
         img_base64 = request.form.get("input_img", "")
         received_file_name = request.form.get("input_file", "")
-        md5_get = hashlib.md5()
-        md5_src = str(time.time()) + str(img_base64)
-        md5_get.update(md5_src.encode("utf-8"))
-
         ext = received_file_name.split(".")[-1]
         if ext == "":
-            return {"result": "未识别的文件类型"}
-        filename = md5_get.hexdigest() + "." + ext
+            return {"result": "Unrecognized file type"}
+        score = time.time()
+        filename = utils.gen_md5(str(time.time()) + str(img_base64)) + "." + ext
         base64_head = img_base64.split(',')[0]
         img_data = base64.b64decode(img_base64.split(',')[-1])
         with open(filename, "wb") as fp:
             fp.write(img_data)
-        input_data = {"image": [filename]}
+        file_list = [filename]
+        if queues_dict[module_name].qsize(
+        ) + 1 > queues_dict[module_name].get_attribute("maxsize"):
+            return {"result": "Too many visitors now, please come back later."}
+        data_2_item(file_list, req_id, score, module_name)
+        data_num = len(file_list)
+        results = []
+        result_len = 0
+        while result_len != data_num:
+            result_len = len(results_dict.get(req_id, []))
+        results = results_dict.get(req_id)
+        results = [i[1] for i in sorted(results, key=lambda k: k[0])]
+        filename = results[0].get("path")
+        ext = filename.split(".")[-1]
+        if filename is not None:
+            output_file = os.path.join("./output", filename)
+            if output_file is not None and os.path.exists(output_file):
+                with open(output_file, "rb") as fp:
+                    output_img_base64 = base64.b64encode(fp.read())
+                os.remove(filename)
+                os.remove(output_file)
+                results = {
+                    "border":
+                    str(results[0]["data"]),
+                    "output_img":
+                    base64_head + "," + str(output_img_base64).replace(
+                        "b'", "").replace("'", "")
+                }
+                return {"result": results}
+        return {"result": str(results)}
 
-        module = ImageModelService.get_module(module_name)
-        method_name = module.desc.attr.map.data['default_signature'].s
-        if method_name != "":
-            predict_method = getattr(module, method_name)
-            print(input_data)
-            try:
-                print("Use gpu is", use_gpu)
-                results = predict_method(data=input_data, use_gpu=use_gpu)
-            except Exception as err:
-                print(err)
-                return {"result": "请检查数据格式!"}
-        else:
-            results = "您调用的模型{}不支持在线预测".format(module_name)
-        os.remove(filename)
-        output_file = os.path.join("./output", filename)
-        if output_file is not None and os.path.exists(output_file):
-            with open(output_file, "rb") as fp:
-                output_img_base64 = base64.b64encode(fp.read())
-            os.remove(output_file)
-        if module.type.startswith("CV"):
-            results = {
-                "border":
-                str(results[0]["data"]),
-                "output_img":
-                base64_head + "," + str(output_img_base64).replace(
-                    "b'", "").replace("'", "")
-            }
-        return {"result": results}
+    def data_2_item(data_list, req_id, score, module_name):
+        global queues_dict
+        for index in range(len(data_list)):
+            queues_dict[module_name].put((score, index, req_id,
+                                          data_list[index]))
 
     @app_instance.route("/predict/text/<module_name>", methods=["POST"])
     def predict_text(module_name):
-        print("123123")
-        print(request.data)
-        print(request.data.get("id"))
-        print(request.form)
-        global use_gpu
-        data = request.form.get("input_text", "")
-        data = data.splitlines()
+        global results_dict, queues_dict
+        req_id = request.data.get("id")
+        data_list = request.form.get("input_text")
+        score = time.time()
+        data_list = data_list.splitlines()
         data_temp = []
-        for index in range(len(data)):
-            data[index] = data[index].strip()
-            if data[index] != "":
-                print("gll")
-                data_temp.append(data[index])
-        data = data_temp
-        if not isinstance(data, list):
-            data = [data]
-        input_data = {"text": data}
+        for index in range(len(data_list)):
+            data_list[index] = data_list[index].strip()
+            if data_list[index] != "":
+                data_temp.append(data_list[index])
+        data_list = data_temp
+        if not isinstance(data_list, list):
+            data_list = [data_list]
+        data_num = len(data_list)
+        if data_num > queues_dict[module_name].get_attribute("maxsize"):
+            return {"result": ["Too much data, please reduce the data."]}
+        if data_num + queues_dict[module_name].qsize(
+        ) > queues_dict[module_name].get_attribute("maxsize"):
+            return {"result": "Too many visitors now, please come back later."}
+        start = time.time()
 
-        module = TextModelService.get_module(module_name)
-        method_name = module.desc.attr.map.data['default_signature'].s
-        if method_name != "":
-            predict_method = getattr(module, method_name)
-            try:
-                print("Use gpu is", use_gpu)
-                results = predict_method(data=input_data, use_gpu=use_gpu)
-            except Exception as err:
-                print(err)
-                return {"result": "请检查数据格式!"}
-        else:
-            results = "您调用的模型{}不支持在线预测".format(module_name)
-
-        if results == []:
-            results = "输入不能为空"
-        return {"result": results}
-
-    @app_instance.route("/predict/<module_name>", methods=["POST"])
-    def predict(module_name):
-        print(request.form)
-        data = request.form.get("input_text", "")
-        img_base64 = request.form.get("input_img", "")
-        received_file_name = request.form.get("input_file", "")
-        filename = None
-        if module_name != "" and data != "":
-            print("ininin")
-            data = data.splitlines()
-            print(data)
-            data_temp = []
-            for index in range(len(data)):
-                data[index] = data[index].strip()
-                if data[index] != "":
-                    print("gll")
-                    data_temp.append(data[index])
-            data = data_temp
-            if not isinstance(data, list):
-                data = [data]
-            input_data = {"text": data}
-        # 如果是图片
-        elif module_name != "" and img_base64 != "":
-            md5_get = hashlib.md5()
-            md5_src = str(time.time()) + str(img_base64)
-            md5_get.update(md5_src.encode("utf-8"))
-            ext = received_file_name.split(".")[-1]
-            if ext == "":
-                return {"result": "未识别的文件类型，请添加后缀"}
-            filename = md5_get.hexdigest() + "." + ext
-            base64_head = img_base64.split(',')[0]
-            img_data = base64.b64decode(img_base64.split(',')[-1])
-            print("123123123")
-            print(img_data)
-            with open(filename, "wb") as fp:
-                fp.write(img_data)
-            input_data = {"image": [filename]}
-        else:
-            return {"result": "不支持的文件类型！"}
-        module = TextModelService.get_module(module_name)
-        method_name = module.desc.attr.map.data['default_signature'].s
-        if method_name != "":
-            predict_method = getattr(module, method_name)
-            print(input_data)
-            try:
-                results = predict_method(data=input_data)
-            except Exception as err:
-                print(err)
-                return {"result": "请检查数据格式!"}
-        else:
-            results = "您调用的模型{}不支持在线预测".format(module_name)
-        output_file = None
-        if filename is not None:
-            os.remove(filename)
-            output_file = os.path.join("./output", filename)
-        print("output_file=", output_file)
-
-        if output_file is not None and os.path.exists(output_file):
-            with open(output_file, "rb") as fp:
-                output_img_base64 = base64.b64encode(fp.read())
-            os.remove(output_file)
-        if module.type.startswith("CV"):
-            results = {
-                "border":
-                str(results[0]["data"]),
-                "output_img":
-                base64_head + "," + str(output_img_base64).replace(
-                    "b'", "").replace("'", "")
-            }
-        print(results)
-        if results == []:
-            results = "输入不能为空."
+        data_2_item(data_list, req_id, score, module_name)
+        print("cunchu need", time.time() - start)
+        results = []
+        result_len = 0
+        start = time.time()
+        while result_len != data_num:
+            result_len = len(results_dict.get(req_id, []))
+        print("cha zhao xuyao", time.time() - start)
+        start = time.time()
+        results = results_dict.get(req_id)
+        results = [i[1] for i in sorted(results, key=lambda k: k[0])]
+        print("sort need", time.time() - start)
         return {"result": results}
 
     return app_instance
 
 
+def config_with_file(configs):
+    global m
+    global nlp_module, cv_module, queues_list, batch_size_list, name_list, \
+        queues_dict, queue_name_list, results_dict
+    m = Manager()
+    m.start()
+    nlp_module = []
+    cv_module = []
+    queues_list = []
+    batch_size_list = []
+    name_list = []
+    queues_dict = {}
+    queue_name_list = []
+    results_dict = mp.Manager().dict()
+    for item in configs:
+        print(item)
+        if item["category"] == "CV":
+            cv_module.append(item["module"])
+        elif item["category"] == "NLP":
+            nlp_module.append(item["module"])
+        queues_list.append(m.get_priorityQueue(maxsize=item["queue_size"]))
+        batch_size_list.append(item["batch_size"])
+        name_list.append(item["module"])
+        queues_dict.update({item["module"]: queues_list[-1]})
+        queue_name_list.append(item["module"])
+
+
 def run(is_use_gpu=False, configs=None):
     global use_gpu
     use_gpu = is_use_gpu
+    if configs is not None:
+        config_with_file(configs)
+    else:
+        print("Start failed cause of missing configuration.")
+        return
     my_app = create_app()
     my_app.run(host="0.0.0.0", port=8888, debug=True)
 
 
 if __name__ == "__main__":
-    run(is_use_gpu=False)
+    configs = [{
+        'category': 'NLP',
+        u'queue_size': 20,
+        u'version': u'1.0.0',
+        u'module': 'lac',
+        u'batch_size': 20
+    },
+               {
+                   'category': 'NLP',
+                   u'queue_size': 20,
+                   u'version': u'1.0.0',
+                   u'module': 'senta_lstm',
+                   u'batch_size': 20
+               },
+               {
+                   'category': 'CV',
+                   u'queue_size': 20,
+                   u'version': u'1.0.0',
+                   u'module': 'yolov3_coco2017',
+                   u'batch_size': 20
+               },
+               {
+                   'category': 'CV',
+                   u'queue_size': 20,
+                   u'version': u'1.0.0',
+                   u'module': 'faster_rcnn_coco2017',
+                   u'batch_size': 20
+               }]
+    run(is_use_gpu=False, configs=configs)
